scripts/requirements.txt
python-libarchive
cryptography==40.0.1

